move interfaces into a new directory
