+++ packages/app/components/Beneficiary-Proposals/BeneficiaryCard.tsx
@@ -2,8 +2,12 @@
 
 import Link from 'next/link';
 import { social } from '../../fixtures/social';
-
-const beneficiaryProposalCard = (beneficiaryProposal) => {
+import {
+  DummyBeneficiaryProposal,
+} from '../../interfaces/beneficiaries';
+const beneficiaryProposalCard = (
+  beneficiaryProposal: DummyBeneficiaryProposal,
+): JSX.Element => {
   return (
     <Link
       key={beneficiaryProposal.name}
