+++ packages/app/components/Beneficiary-Proposals/ChallengePeriodVoting.tsx
@@ -2,7 +2,7 @@ import { useContext, useEffect, useState } from 'react';
 import { store } from 'app/store';
 
 import { setDualActionModal } from '../../app/actions';
-import { DummyBeneficiaryProposal } from '../../pages/beneficiary-proposals/interfaces';
+import {DummyBeneficiaryProposal} from '../../interfaces/beneficiaries'
 import CurrentStandings from './CurrentStandings';
 
 // TODO: Add types
