+++ packages/app/components/Beneficiary-Proposals/CompletedVoting.tsx
@@ -1,9 +1,5 @@
-import { useContext, useEffect, useState } from 'react';
-import { store } from 'app/store';
-
-import { DummyBeneficiaryProposal } from '../../pages/beneficiary-proposals/interfaces';
+import { DummyBeneficiaryProposal } from '../../interfaces/beneficiaries';
 import CurrentStandings from './CurrentStandings';
-import { beneficiaryProposalFixture } from 'fixtures/beneficiaryProposals';
 
 // TODO: Add types
 const CompletedVoting = (beneficiaryProposal: DummyBeneficiaryProposal) => {
