+++ packages/app/interfaces/beneficiaries.tsx
@@ -22,6 +22,5 @@ export interface DummyBeneficiaryProposal {
   votesAgainst: number;
   currentStage: Stage;
   stageDeadline: Date;
-  proofOfOwnership: string;
-  
+  proofOfOwnership: string;  
 }
